+++ src/registers/reg_macro.rs
@@ -1,5 +1,5 @@
-pub use crate::utils::create_mask;
 pub use paste::paste;
+// use utils::create_mask;
 
 pub trait Register {
     const ADDRESS: u8;
