+++ tests/integration_test.rs
@@ -2,8 +2,8 @@ mod stub;
 
 use embedded_hal::spi::ErrorKind;
 use stub::*;
+use tmc5160_driver::registers::VMAX;
 use tmc5160_driver::Tmc5160;
-use tmc5160_driver::VMAX;
 
 type Driver = Tmc5160<StubSpiDevice>;
